Events removed when course is removed + Event Colors
When a course is removed all events on the calendar pertaining to that
course are removed. Each course now has a cooler assigned to it when
added to the model. The data table highlights selected course sections
with a color that matches the events colour on the calendar for better
interaction feedback.